--- a/.travis.yml
+++ b/.travis.yml
@@ -3,7 +3,7 @@ python:
   - "2.7"
   - "3.3"
   - "3.4"
-install: "pip install -r requirements.pip --use-mirrors"
+install: "pip install -r requirements_py${TRAVIS_PYTHON_VERSION%.*}.pip"
 script: python setup.py test
 notifications:
     email: