--- a/requirements_py2.pip
+++ b/requirements_py2.pip
@@ -1,0 +1,2 @@
+-r requirements_py3.pip
+backports.functools_lru_cache==1.0.1