--- a/tox.ini
+++ b/tox.ini
@@ -6,4 +6,5 @@ setenv =
     PYTHONPATH = {toxinidir}:{toxinidir}/colorific
 commands = python setup.py test
 deps =
-    -r{toxinidir}/requirements.pip
+    py3{3,4}: -r{toxinidir}/requirements_py3.pip
+    py27: -r{toxinidir}/requirements_py2.pip