@larsyencken 2015-05-29
Good work! I'll see that this gets merged, a little admin to do on my side first.