@dhotson 2015-06-02
Sorry about the wait. Thanks for the PR! :)